+++ ExtraTerminalCommands/CHANGELOG.md
@@ -1,4 +1,6 @@
-# V1.3.5
+# V1.4.0
+## Additions:
+- 's' command, does the same as the vanilla 'switch' command. Includes player
 
-## Bug Fixes:
-- Weather Filter condition broken
+## Known Bugs:
+- Will say that the 's' command doesn't exist whilst it does, might be fixed in the future